refactor: docs/js를 ES Modules로 목적별 분할
dashboard.js(439줄)를 4개 모듈로 분리:
- main.js: 진입점, 모드 전환, 데이터 로딩 오케스트레이션
- charts.js: Chart.js 차트 3개 (performance, strategy, allocation)
- ui.js: DOM 업데이트 (카드, 결정로직, 거래내역)
- utils.js: 공용 헬퍼 (getRegimeColorClass)

개선사항:
- ES Modules import/export로 명시적 의존성 관리
- 순차 fetch를 Promise.all로 병렬화
- 미사용 renderAllocationHistory() 함수 제거
