+++ docs/index.html
@@ -194,7 +194,7 @@
     <!-- Chart.js & Bootstrap JS -->
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <!-- Dashboard Logic -->
-    <script src="js/dashboard.js"></script>
+    <!-- Dashboard Logic (ES Modules) -->
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
